#8 refactor parameters code
