+++ src/main/java/com/alchemyapi/api/AlchemyAPI.java
@@ -1,25 +1,39 @@
 package com.alchemyapi.api;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.InputStreamReader;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.FileNotFoundException;
-
-import java.net.HttpURLConnection;
-import java.net.URL;
+import com.alchemyapi.api.parameters.CategoryParameters;
+import com.alchemyapi.api.parameters.CombinedParameters;
+import com.alchemyapi.api.parameters.ConceptParameters;
+import com.alchemyapi.api.parameters.ConstraintQueryParameters;
+import com.alchemyapi.api.parameters.ImageParameters;
+import com.alchemyapi.api.parameters.KeywordParameters;
+import com.alchemyapi.api.parameters.LanguageParameters;
+import com.alchemyapi.api.parameters.NamedEntityParameters;
+import com.alchemyapi.api.parameters.Parameters;
+import com.alchemyapi.api.parameters.RelationParameters;
+import com.alchemyapi.api.parameters.TargetedSentimentParameters;
+import com.alchemyapi.api.parameters.TaxonomyParameters;
+import com.alchemyapi.api.parameters.TextParameters;
+import org.w3c.dom.Document;
+import org.w3c.dom.NodeList;
+import org.xml.sax.SAXException;
 
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.xpath.*;
-
-import org.w3c.dom.Document;
-import org.w3c.dom.NodeList;
-
-import org.xml.sax.SAXException;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathExpression;
+import javax.xml.xpath.XPathExpressionException;
+import javax.xml.xpath.XPathFactory;
+import java.io.BufferedReader;
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
 
 public class AlchemyAPI {
 
@@ -84,10 +98,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetAuthor(url, new AlchemyAPI_Params());
+        return URLGetAuthor(url, new Parameters());
     }
 
-    public Document URLGetAuthor(String url, AlchemyAPI_Params params)
+    public Document URLGetAuthor(String url, Parameters params)
         throws IOException, SAXException, ParserConfigurationException,
                XPathExpressionException
     {
@@ -102,11 +116,11 @@ public class AlchemyAPI {
         throws IOException, SAXException, ParserConfigurationException,
                XPathExpressionException
     {
-        return HTMLGetAuthor(html, url, new AlchemyAPI_Params());
+        return HTMLGetAuthor(html, url, new Parameters());
     }
 
     public Document HTMLGetAuthor(String html, String url,
-			          AlchemyAPI_Params params)
+			          Parameters params)
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
@@ -122,10 +136,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetRankedNamedEntities(url, new AlchemyAPI_NamedEntityParams());
+        return URLGetRankedNamedEntities(url, new NamedEntityParameters());
     }
     
-    public Document URLGetRankedNamedEntities(String url, AlchemyAPI_NamedEntityParams params)
+    public Document URLGetRankedNamedEntities(String url, NamedEntityParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
     {
@@ -140,11 +154,11 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetRankedNamedEntities(html, url, new AlchemyAPI_NamedEntityParams());
+        return HTMLGetRankedNamedEntities(html, url, new NamedEntityParameters());
     }
     
 
-    public Document HTMLGetRankedNamedEntities(String html, String url, AlchemyAPI_NamedEntityParams params)
+    public Document HTMLGetRankedNamedEntities(String html, String url, NamedEntityParameters params)
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
@@ -160,10 +174,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return TextGetRankedNamedEntities(text, new AlchemyAPI_NamedEntityParams());
+        return TextGetRankedNamedEntities(text, new NamedEntityParameters());
     }
     
-    public Document TextGetRankedNamedEntities(String text, AlchemyAPI_NamedEntityParams params)
+    public Document TextGetRankedNamedEntities(String text, NamedEntityParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
     {
@@ -180,10 +194,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetRankedConcepts(url, new AlchemyAPI_ConceptParams());
+        return URLGetRankedConcepts(url, new ConceptParameters());
     }
     
-    public Document URLGetRankedConcepts(String url, AlchemyAPI_ConceptParams params)
+    public Document URLGetRankedConcepts(String url, ConceptParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
     {
@@ -199,10 +213,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetRankedConcepts(html, url, new AlchemyAPI_ConceptParams());
+        return HTMLGetRankedConcepts(html, url, new ConceptParameters());
     }
     
-    public Document HTMLGetRankedConcepts(String html, String url, AlchemyAPI_ConceptParams params)
+    public Document HTMLGetRankedConcepts(String html, String url, ConceptParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -216,10 +230,10 @@ public class AlchemyAPI {
 
     public Document TextGetRankedConcepts(String text) throws IOException, SAXException,
             ParserConfigurationException, XPathExpressionException {
-        return TextGetRankedConcepts(text, new AlchemyAPI_ConceptParams());
+        return TextGetRankedConcepts(text, new ConceptParameters());
     }
     
-    public Document TextGetRankedConcepts(String text, AlchemyAPI_ConceptParams params) throws IOException, SAXException,
+    public Document TextGetRankedConcepts(String text, ConceptParameters params) throws IOException, SAXException,
     ParserConfigurationException, XPathExpressionException 
     {
 		CheckText(text);
@@ -235,10 +249,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetRankedKeywords(url, new AlchemyAPI_KeywordParams());
+        return URLGetRankedKeywords(url, new KeywordParameters());
     }
     
-    public Document URLGetRankedKeywords(String url, AlchemyAPI_KeywordParams params)
+    public Document URLGetRankedKeywords(String url, KeywordParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
     {
@@ -254,10 +268,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetRankedKeywords(html, url, new AlchemyAPI_KeywordParams());
+        return HTMLGetRankedKeywords(html, url, new KeywordParameters());
     }
     
-    public Document HTMLGetRankedKeywords(String html, String url, AlchemyAPI_KeywordParams params)
+    public Document HTMLGetRankedKeywords(String html, String url, KeywordParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -271,10 +285,10 @@ public class AlchemyAPI {
 
     public Document TextGetRankedKeywords(String text) throws IOException, SAXException,
             ParserConfigurationException, XPathExpressionException {
-        return TextGetRankedKeywords(text, new AlchemyAPI_KeywordParams());
+        return TextGetRankedKeywords(text, new KeywordParameters());
     }
     
-    public Document TextGetRankedKeywords(String text, AlchemyAPI_KeywordParams params) throws IOException, SAXException,
+    public Document TextGetRankedKeywords(String text, KeywordParameters params) throws IOException, SAXException,
     ParserConfigurationException, XPathExpressionException 
     {
 		CheckText(text);
@@ -288,10 +302,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetLanguage(url, new AlchemyAPI_LanguageParams());
+        return URLGetLanguage(url, new LanguageParameters());
     }
     
-    public Document URLGetLanguage(String url, AlchemyAPI_LanguageParams params)
+    public Document URLGetLanguage(String url, LanguageParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -306,10 +320,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetLanguage(html, url, new AlchemyAPI_LanguageParams());
+        return HTMLGetLanguage(html, url, new LanguageParameters());
     }
     
-    public Document HTMLGetLanguage(String html, String url, AlchemyAPI_LanguageParams params)
+    public Document HTMLGetLanguage(String html, String url, LanguageParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -325,10 +339,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return TextGetLanguage(text, new AlchemyAPI_LanguageParams());
+        return TextGetLanguage(text, new LanguageParameters());
     }
     
-    public Document TextGetLanguage(String text, AlchemyAPI_LanguageParams params)
+    public Document TextGetLanguage(String text, LanguageParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -343,10 +357,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetCategory(url, new AlchemyAPI_CategoryParams());
+        return URLGetCategory(url, new CategoryParameters());
     }
     
-    public Document URLGetCategory(String url, AlchemyAPI_CategoryParams params)
+    public Document URLGetCategory(String url, CategoryParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -361,10 +375,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetCategory(html, url, new AlchemyAPI_CategoryParams());
+        return HTMLGetCategory(html, url, new CategoryParameters());
     }
     
-    public Document HTMLGetCategory(String html, String url, AlchemyAPI_CategoryParams params)
+    public Document HTMLGetCategory(String html, String url, CategoryParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -380,10 +394,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return TextGetCategory(text, new AlchemyAPI_TextParams());
+        return TextGetCategory(text, new TextParameters());
     }
     
-    public Document TextGetCategory(String text, AlchemyAPI_TextParams params)
+    public Document TextGetCategory(String text, TextParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -398,10 +412,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetText(url, new AlchemyAPI_TextParams());
+        return URLGetText(url, new TextParameters());
     }
     
-    public Document URLGetText(String url, AlchemyAPI_TextParams params)
+    public Document URLGetText(String url, TextParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -416,10 +430,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetText(html, url, new AlchemyAPI_TextParams());
+        return HTMLGetText(html, url, new TextParameters());
     }
     
-    public Document HTMLGetText(String html, String url, AlchemyAPI_TextParams params)
+    public Document HTMLGetText(String html, String url, TextParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -435,10 +449,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetRawText(url, new AlchemyAPI_Params());
+        return URLGetRawText(url, new Parameters());
     }
     
-    public Document URLGetRawText(String url, AlchemyAPI_Params params)
+    public Document URLGetRawText(String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -453,10 +467,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetRawText(html, url, new AlchemyAPI_Params());
+        return HTMLGetRawText(html, url, new Parameters());
     }
     
-    public Document HTMLGetRawText(String html, String url, AlchemyAPI_Params params)
+    public Document HTMLGetRawText(String html, String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -472,10 +486,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetTitle(url, new AlchemyAPI_Params());
+        return URLGetTitle(url, new Parameters());
     }
     
-    public Document URLGetTitle(String url, AlchemyAPI_Params params)
+    public Document URLGetTitle(String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -490,10 +504,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetTitle(html, url, new AlchemyAPI_Params());
+        return HTMLGetTitle(html, url, new Parameters());
     }
     
-    public Document HTMLGetTitle(String html, String url, AlchemyAPI_Params params)
+    public Document HTMLGetTitle(String html, String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -509,10 +523,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetFeedLinks(url, new AlchemyAPI_Params());
+        return URLGetFeedLinks(url, new Parameters());
     }
     
-    public Document URLGetFeedLinks(String url, AlchemyAPI_Params params)
+    public Document URLGetFeedLinks(String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -527,10 +541,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetFeedLinks(html, url, new AlchemyAPI_Params());
+        return HTMLGetFeedLinks(html, url, new Parameters());
     }
     
-    public Document HTMLGetFeedLinks(String html, String url, AlchemyAPI_Params params)
+    public Document HTMLGetFeedLinks(String html, String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -546,10 +560,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetMicroformats(url, new AlchemyAPI_Params());
+        return URLGetMicroformats(url, new Parameters());
     }
     
-    public Document URLGetMicroformats(String url, AlchemyAPI_Params params)
+    public Document URLGetMicroformats(String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -564,10 +578,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetMicroformats(html, url, new AlchemyAPI_Params());
+        return HTMLGetMicroformats(html, url, new Parameters());
     }
     
-    public Document HTMLGetMicroformats(String html, String url, AlchemyAPI_Params params)
+    public Document HTMLGetMicroformats(String html, String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -583,10 +597,10 @@ public class AlchemyAPI {
         throws IOException, XPathExpressionException,
                SAXException, ParserConfigurationException
     {
-        return URLGetConstraintQuery(url, query, new AlchemyAPI_ConstraintQueryParams());
+        return URLGetConstraintQuery(url, query, new ConstraintQueryParameters());
     }
     
-    public Document URLGetConstraintQuery(String url, String query, AlchemyAPI_ConstraintQueryParams params)
+    public Document URLGetConstraintQuery(String url, String query, ConstraintQueryParameters params)
     throws IOException, XPathExpressionException,
            SAXException, ParserConfigurationException
 	{
@@ -605,10 +619,10 @@ public class AlchemyAPI {
         throws IOException, XPathExpressionException,
                SAXException, ParserConfigurationException
     {
-        return HTMLGetConstraintQuery(html, url, query, new AlchemyAPI_ConstraintQueryParams());
+        return HTMLGetConstraintQuery(html, url, query, new ConstraintQueryParameters());
     }
     
-    public Document HTMLGetConstraintQuery(String html, String url, String query, AlchemyAPI_ConstraintQueryParams params)
+    public Document HTMLGetConstraintQuery(String html, String url, String query, ConstraintQueryParameters params)
     throws IOException, XPathExpressionException,
            SAXException, ParserConfigurationException
 	{
@@ -627,10 +641,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetTextSentiment(url, new AlchemyAPI_Params());
+        return URLGetTextSentiment(url, new Parameters());
     }
     
-    public Document URLGetTextSentiment(String url, AlchemyAPI_Params params)
+    public Document URLGetTextSentiment(String url, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
     {
@@ -646,10 +660,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetTextSentiment(html, url, new AlchemyAPI_Params());
+        return HTMLGetTextSentiment(html, url, new Parameters());
     }
     
-    public Document HTMLGetTextSentiment(String html, String url, AlchemyAPI_Params params)
+    public Document HTMLGetTextSentiment(String html, String url, Parameters params)
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
@@ -665,10 +679,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return TextGetTextSentiment(text, new AlchemyAPI_Params());
+        return TextGetTextSentiment(text, new Parameters());
     }
     
-    public Document TextGetTextSentiment(String text, AlchemyAPI_Params params)
+    public Document TextGetTextSentiment(String text, Parameters params)
         throws IOException, SAXException,
         ParserConfigurationException, XPathExpressionException 
     {
@@ -686,11 +700,11 @@ public class AlchemyAPI {
                XPathExpressionException
     {
         return URLGetTargetedSentiment(url, target,
-                                       new AlchemyAPI_TargetedSentimentParams());
+                                       new TargetedSentimentParameters());
     }
 
     public Document URLGetTargetedSentiment(String url, String target,
-                                            AlchemyAPI_TargetedSentimentParams params)
+                                            TargetedSentimentParameters params)
         throws IOException, SAXException, ParserConfigurationException,
                XPathExpressionException
     {
@@ -708,11 +722,11 @@ public class AlchemyAPI {
                XPathExpressionException
     {
         return HTMLGetTargetedSentiment(html, url, target,
-                                        new AlchemyAPI_TargetedSentimentParams());
+                                        new TargetedSentimentParameters());
     }
 
     public Document HTMLGetTargetedSentiment(String html, String url, String target,
-                                             AlchemyAPI_TargetedSentimentParams params)
+                                             TargetedSentimentParameters params)
         throws IOException, SAXException, ParserConfigurationException,
                XPathExpressionException
     {
@@ -731,11 +745,11 @@ public class AlchemyAPI {
                XPathExpressionException
     {
         return TextGetTargetedSentiment(text, target,
-                                        new AlchemyAPI_TargetedSentimentParams());
+                                        new TargetedSentimentParameters());
     }
 
     public Document TextGetTargetedSentiment(String text, String target,
-                                             AlchemyAPI_TargetedSentimentParams params)
+                                             TargetedSentimentParameters params)
         throws IOException, SAXException, ParserConfigurationException,
 	       XPathExpressionException
     {
@@ -753,10 +767,10 @@ public class AlchemyAPI {
     public Document URLGetRelations(String url) throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetRelations(url, new AlchemyAPI_RelationParams());
+        return URLGetRelations(url, new RelationParameters());
     }
     
-    public Document URLGetRelations(String url, AlchemyAPI_RelationParams params)
+    public Document URLGetRelations(String url, RelationParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -771,10 +785,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetRelations(html, url, new AlchemyAPI_RelationParams());
+        return HTMLGetRelations(html, url, new RelationParameters());
     }
     
-    public Document HTMLGetRelations(String html, String url, AlchemyAPI_RelationParams params)
+    public Document HTMLGetRelations(String html, String url, RelationParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -790,10 +804,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return TextGetRelations(text, new AlchemyAPI_RelationParams());
+        return TextGetRelations(text, new RelationParameters());
     }
     
-    public Document TextGetRelations(String text, AlchemyAPI_RelationParams params)
+    public Document TextGetRelations(String text, RelationParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -809,12 +823,12 @@ public class AlchemyAPI {
     public Document URLGetCombined(String url) throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-	AlchemyAPI_CombinedParams params = new AlchemyAPI_CombinedParams();
+	CombinedParameters params = new CombinedParameters();
 	params.setExtractAll();
         return URLGetCombined(url, params);
     }
     
-    public Document URLGetCombined(String url, AlchemyAPI_CombinedParams params)
+    public Document URLGetCombined(String url, CombinedParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -829,12 +843,12 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-	AlchemyAPI_CombinedParams params = new AlchemyAPI_CombinedParams();
+	CombinedParameters params = new CombinedParameters();
 	params.setExtractAll();
         return TextGetCombined(text, params);
     }
     
-    public Document TextGetCombined(String text, AlchemyAPI_CombinedParams params)
+    public Document TextGetCombined(String text, CombinedParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -850,10 +864,10 @@ public class AlchemyAPI {
     public Document URLGetTaxonomy(String url) throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetTaxonomy(url, new AlchemyAPI_TaxonomyParams());
+        return URLGetTaxonomy(url, new TaxonomyParameters());
     }
     
-    public Document URLGetTaxonomy(String url, AlchemyAPI_TaxonomyParams params)
+    public Document URLGetTaxonomy(String url, TaxonomyParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -868,10 +882,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return HTMLGetTaxonomy(html, url, new AlchemyAPI_TaxonomyParams());
+        return HTMLGetTaxonomy(html, url, new TaxonomyParameters());
     }
     
-    public Document HTMLGetTaxonomy(String html, String url, AlchemyAPI_TaxonomyParams params)
+    public Document HTMLGetTaxonomy(String html, String url, TaxonomyParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -887,10 +901,10 @@ public class AlchemyAPI {
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return TextGetTaxonomy(text, new AlchemyAPI_TaxonomyParams());
+        return TextGetTaxonomy(text, new TaxonomyParameters());
     }
     
-    public Document TextGetTaxonomy(String text, AlchemyAPI_TaxonomyParams params)
+    public Document TextGetTaxonomy(String text, TaxonomyParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -906,10 +920,10 @@ public class AlchemyAPI {
     public Document URLGetImage(String url) throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetImage(url, new AlchemyAPI_ImageParams());
+        return URLGetImage(url, new ImageParameters());
     }
     
-    public Document URLGetImage(String url, AlchemyAPI_ImageParams params)
+    public Document URLGetImage(String url, ImageParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -923,10 +937,10 @@ public class AlchemyAPI {
     public Document URLGetRankedImageKeywords(String url) throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
-        return URLGetRankedImageKeywords(url, new AlchemyAPI_ImageParams());
+        return URLGetRankedImageKeywords(url, new ImageParameters());
     }
     
-    public Document URLGetRankedImageKeywords(String url, AlchemyAPI_ImageParams params)
+    public Document URLGetRankedImageKeywords(String url, ImageParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
     {
@@ -937,7 +951,7 @@ public class AlchemyAPI {
         return GET("URLGetRankedImageKeywords", "url", params);
     }
 
-    public Document ImageGetRankedImageKeywords(AlchemyAPI_ImageParams params)
+    public Document ImageGetRankedImageKeywords(ImageParameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
     {
@@ -976,7 +990,7 @@ public class AlchemyAPI {
             throw new IllegalArgumentException("Enter an URL to analyze.");
     }
     
-    private Document GET(String callName, String callPrefix, AlchemyAPI_Params params)
+    private Document GET(String callName, String callPrefix, Parameters params)
     throws IOException, SAXException,
            ParserConfigurationException, XPathExpressionException
 	{
@@ -992,7 +1006,7 @@ public class AlchemyAPI {
 	    return doRequest(handle, params.getOutputMode());
 	}
 
-    private Document POST(String callName, String callPrefix, AlchemyAPI_Params params)
+    private Document POST(String callName, String callPrefix, Parameters params)
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
     {
@@ -1015,6 +1029,8 @@ public class AlchemyAPI {
         return doRequest(handle, params.getOutputMode());
     }
 
+    // TODO currently hard-coded to handle xml, create xml/json adapter based on request
+    // TODO return pojo with parsed field, but allow a "raw" xml/json getter to protect against api updates
     private Document doRequest(HttpURLConnection handle, String outputMode)
         throws IOException, SAXException,
                ParserConfigurationException, XPathExpressionException
@@ -1027,7 +1043,7 @@ public class AlchemyAPI {
 
         XPathFactory factory = XPathFactory.newInstance();
 
-        if(AlchemyAPI_Params.OUTPUT_XML.equals(outputMode)) {
+        if(Parameters.OUTPUT_XML.equals(outputMode)) {
         	String statusStr = getNodeValue(factory, doc, "/results/status/text()");
         	if (null == statusStr || !statusStr.equals("OK")) {
         		String statusInfoStr = getNodeValue(factory, doc, "/results/statusInfo/text()");
@@ -1037,7 +1053,7 @@ public class AlchemyAPI {
         		throw new IOException("Error making API call: " + statusStr + '.');
         	}
         }
-        else if(AlchemyAPI_Params.OUTPUT_RDF.equals(outputMode)) {
+        else if(Parameters.OUTPUT_RDF.equals(outputMode)) {
         	String statusStr = getNodeValue(factory, doc, "//RDF/Description/ResultStatus/text()");
         	if (null == statusStr || !statusStr.equals("OK")) {
         		String statusInfoStr = getNodeValue(factory, doc, "//RDF/Description/ResultStatus/text()");
